Modifies most files to use absolute imports. Moves most of the widgets from BaseProcess into a new custom_widget.py Adds modern typing to binary.py
